+++ packages/core/src/admin-dashboard.server.ts
@@ -3,8 +3,7 @@ export { sendOneTimeCode } from './modules/authentication/use-cases/send-one-tim
 export { verifyOneTimeCode } from './modules/authentication/use-cases/verify-one-time-code';
 export { getEvent } from './modules/events/queries/get-event';
 export { listEvents } from './modules/events/queries/list-events';
-export { addEventRecordingLink } from './modules/events/use-cases/add-event-recording-link';
-export { createEvent } from './modules/events/use-cases/create-event';
+export { addEventRecordingLink, createEvent } from './modules/events/events';
 export { getFeatureFlag } from './modules/feature-flags/queries/get-feature-flag';
 export { listFeatureFlags } from './modules/feature-flags/queries/list-feature-flags';
 export { createFeatureFlag } from './modules/feature-flags/use-cases/create-feature-flag';
