+++ packages/core/src/member-profile.server.ts
@@ -19,11 +19,13 @@ export { searchCrunchbaseOrganizations } from './modules/employment/queries/sear
 export { addWorkExperience } from './modules/employment/use-cases/add-work-experience';
 export { deleteWorkExperience } from './modules/employment/use-cases/delete-work-experience';
 export { editWorkExperience } from './modules/employment/use-cases/edit-work-experience';
-export { countEventAttendees } from './modules/events/queries/count-event-attendees';
+export {
+  countEventAttendees,
+  listEventAttendees,
+} from './modules/events/event-attendees';
 export { countPastEvents } from './modules/events/queries/count-past-events';
 export { countUpcomingEvents } from './modules/events/queries/count-upcoming-events';
 export { getEvent } from './modules/events/queries/get-event';
-export { listEventAttendees } from './modules/events/queries/list-event-attendees';
 export { isFeatureFlagEnabled } from './modules/feature-flags/queries/is-feature-flag-enabled';
 export { getIcebreakerPrompts } from './modules/icebreakers/queries/get-icebreaker-prompts';
 export { getIcebreakerResponses } from './modules/icebreakers/queries/get-icebreaker-responses';
