+++ packages/core/src/modules/events/event-registrations.ts
@@ -1,9 +1,9 @@
 import { db } from '@oyster/db';
-import type { Gender } from '@oyster/types';
-import { FORMATTED_GENDER } from '@oyster/types';
+import { FORMATTED_GENDER, type Gender } from '@oyster/types';
 
 import { job } from '@/infrastructure/bull';
 import { type GetBullJobData } from '@/infrastructure/bull.types';
+import { registerForAirmeetEvent } from './airmeet';
 
 export async function onRegisteredForEvent({
   eventId,
@@ -43,5 +43,28 @@ export async function onRegisteredForEvent({
       'Member #': student.number,
       'Registered On': new Date().toISOString(),
     },
+  });
+}
+
+export async function registerForEvent({
+  eventId,
+  studentId,
+}: GetBullJobData<'event.register'>) {
+  const student = await db
+    .selectFrom('students')
+    .select(['email', 'firstName', 'lastName'])
+    .where('id', '=', studentId)
+    .executeTakeFirstOrThrow();
+
+  await registerForAirmeetEvent({
+    email: student.email,
+    eventId,
+    firstName: student.firstName,
+    lastName: student.lastName,
+  });
+
+  job('event.registered', {
+    eventId,
+    studentId,
   });
 }
