--- packages/core/src/modules/events/events.types.ts
@@ -1,13 +1,0 @@
-import { type z } from 'zod';
-
-import { Event } from '@oyster/types';
-
-// Use Cases
-
-export const AddEventRecordingLinkInput = Event.pick({
-  recordingLink: true,
-});
-
-export type AddEventRecordingLinkInput = z.infer<
-  typeof AddEventRecordingLinkInput
->;
