--- packages/core/src/modules/events/events/event-attended.ts
@@ -1,18 +1,0 @@
-import { job } from '@/infrastructure/bull';
-import { type GetBullJobData } from '@/infrastructure/bull.types';
-
-export async function onEventAttended({
-  eventId,
-  studentId,
-}: GetBullJobData<'event.attended'>) {
-  job('student.activation_requirement_completed', {
-    requirement: 'attend_event',
-    studentId,
-  });
-
-  job('gamification.activity.completed', {
-    eventId,
-    studentId,
-    type: 'attend_event',
-  });
-}
